+++ scripts/protocolFilters/index.ts
@@ -17,7 +17,7 @@ import { filter as filterCeloTransactions } from './celoTransactions'
 import { filter as filterRhino } from './rhino'
 import { filter as filterScoutGameV0 } from './scoutGameV0'
 import { filter as filterLiskV0 } from './lisk-v0'
-import { filter as filterBlazeTheBase } from './blaze-the-base'
+import { filter as filterBaseV0 } from './baseV0'
 
 export const protocolFilters: Record<Protocol, FilterFn> = {
   beefy: _createFilter(filterBeefy),
@@ -32,7 +32,7 @@ export const protocolFilters: Record<Protocol, FilterFn> = {
   rhino: _createFilter(filterRhino),
   'scout-game-v0': _createFilter(filterScoutGameV0),
   'lisk-v0': _createFilter(filterLiskV0),
-  'blaze-the-base': _createFilter(filterBlazeTheBase),
+  'base-v0': _createFilter(filterBaseV0),
 }
 
 const BATCH_SIZE = 20
